refactor: edit comments pagination
