+++ src/components/Post/PostComments/components/Comments/Comments.tsx
@@ -1,4 +1,4 @@
-import { FC, useEffect, useRef } from "react";
+import { FC, useEffect } from "react";
 import cn from "classnames";
 
 import { CommentsProps } from "./Comments.type";
@@ -23,28 +23,6 @@ const Comments: FC<CommentsProps> = ({
   setNickName,
   setReplyToUserId,
 }) => {
-  const startLoaderRef = useRef(null);
-
-  useEffect(() => {
-    const loader = startLoaderRef.current;
-    const observer = new IntersectionObserver(([entries]) => {
-      if (entries.isIntersecting) {
-        setPage(1);
-      }
-    });
-
-    if (loader) {
-      observer.observe(loader);
-    }
-
-    return () => {
-      if (loader) {
-        observer.unobserve(loader);
-      }
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [startLoaderRef.current]);
-
   useEffect(() => {
     comments.forEach((el) => {
       const observer = new IntersectionObserver(([entries]) => {
@@ -83,27 +61,22 @@ const Comments: FC<CommentsProps> = ({
     const classNames = itemClassNames(counter);
 
     return (
-      <>
-        {!counter ? (
-          <div ref={startLoaderRef} className="invisible"></div>
-        ) : null}
-        <ul className={styles["feed__list"]}>
-          {sortedComments.map((i) => {
-            return (
-              <li key={i.id} className={classNames}>
-                <Comment
-                  {...i}
-                  setId={setId}
-                  setNickName={setNickName}
-                  setReplyToUserId={setReplyToUserId}
-                  postId={postId}
-                />
-                {i.comments?.length ? renderComments(i.comments, 1) : null}
-              </li>
-            );
-          })}
-        </ul>
-      </>
+      <ul className={styles["feed__list"]}>
+        {sortedComments.map((i) => {
+          return (
+            <li key={i.id} className={classNames}>
+              <Comment
+                {...i}
+                setId={setId}
+                setNickName={setNickName}
+                setReplyToUserId={setReplyToUserId}
+                postId={postId}
+              />
+              {i.comments?.length ? renderComments(i.comments, 1) : null}
+            </li>
+          );
+        })}
+      </ul>
     );
   };
